added edit-post and add-post components in post

+++ svt-backend/src/main/java/rs/ac/uns/ftn/svtkvtproject/security/TokenUtils.java
@@ -79,7 +79,7 @@ public class TokenUtils {
                 .signWith(SignatureAlgorithm.HS512, secret).compact();
     }
 
-    public int getExpiredIn() {
-        return expiration.intValue();
+    public Long getExpiredIn() {
+        return expiration;
     }
 }
